_read_record: Rename to '_parse_damon_result'
This commit renames '_rad_record' to '_parse_damon_result' for more
general use cases.
